manus-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.12
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
lxml==5.4.0
multidict==6.4.4
openai==1.82.1
packaging==25.0
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
